Refactor attention mechanism to use half-precision (FP16) for weights and data
- Changed data types for attention weights, gradients, and buffers from float to half.
- Updated memory allocation and data copying in training loop to accommodate half-precision.
- Adjusted forward and backward pass function signatures to accept half-precision data.
- Modified loss computation to maintain float precision for accumulation.

+++ data.h
@@ -4,9 +4,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
-#include <cblas.h>
+#include <cuda_fp16.h>
 
-void generate_data(float** X, float** y, int seq_len, int num_samples, int d_model, float range_min, float range_max);
-void save_data(float* X, float* y, int seq_len, int num_samples, int d_model, const char* filename);
+void generate_data(half** X, half** y, int seq_len, int num_samples, int d_model, float range_min, float range_max);
+void save_data(half* X, half* y, int seq_len, int num_samples, int d_model, const char* filename);
 
 #endif
